resistors: Fixes to evaluation, save app run config

+++ backends/resistors/src/main/kotlin/org/njsoly/tinker/resistors/domain/ResistorDetails.kt
@@ -1,7 +1,9 @@
 package org.njsoly.tinker.resistors.domain
 
+import java.math.BigDecimal
+
 class ResistorDetails(
-    val value: Double,
+    val value: BigDecimal,
     val engineeringNotation: String,
     val tolerance: Double? = null,
     val bandPattern: ResistorBandPattern
